done modal on delete board

+++ workspaces/views/board_views/delete_board.py
@@ -1,25 +1,38 @@
 from .importer import *
 
+@require_http_methods(["POST", "GET"])
 @login_required
-@require_http_methods(['POST'])
 def delete_board(request: HttpRequest) -> HttpResponse:
-  try:
-    encrypted_board_id = request.POST.get('board_id')
-    encrypted_workspace_id = request.POST.get('workspace_id')
-
-    board_id = get_model_id(encrypted_board_id)
-    workspace_id = get_model_id(encrypted_workspace_id)
-    
-    board = Board.objects.filter(Q(id=board_id) & Q(workspace_id=workspace_id))
-
-    if board.exists():
-      board.delete()
-      messages.success(request, "Successfully to delete a board")
-      return redirect(urls.reverse('boards', args=[encrypted_workspace_id]))
-    
-    messages.warning(request, Const.NOT_FOUND_BOARD)
   
-  except Exception as err:
-    exception_message_dispatcher(request, err)
-  
-  return redirect(urls.reverse('workspaces'))
+  if request.method == 'GET':
+    encrypted_id = request.GET.get('id')
+    id = get_model_id(encrypted_id)
+    dialog_confirmation = "Are you sure to delete this board? " + \
+                          "you might lose all of cards you've created and there is no way to get them back!"
+
+    context = {
+      "model_id": id,
+      "dialog_confirmation": dialog_confirmation,
+      "model_name": "board" 
+    }
+    return render(request, "confirmation.html", context)
+
+  else:
+    try:
+      encrypted_board_id = request.POST.get('id')
+      board_id = get_model_id(encrypted_board_id)
+      
+      board = Board.objects.filter(Q(id=board_id))
+
+      if board.exists():
+        board.delete()
+        messages.success(request, "Successfully to delete a board")
+        return djrender(request, "messages.html")
+      
+      messages.warning(request, Const.NOT_FOUND_BOARD)
+
+    except Exception as err:
+      exception_message_dispatcher(request, err)
+    
+    return redirect(urls.reverse('workspaces'))
+
